requirements.txt
Flask==3.0.2
eth-account==0.5.6
web3==5.28.0
mnemonic==0.20
bip32utils==0.3.post4
bip44==0.1.0
pymongo==4.0.1
bson==0.5.10
flask_mongoengine

